refactor: definitions widget and list
