+++ src/2_widgets/Definitions/DefinitionsList/item/DefinitionItem.module.scss
@@ -1,5 +1,5 @@
-@import '@shared/styles/colors';
-@import '@shared/styles/mixins';
+@import '@styles/colors';
+@import '@styles/mixins/font';
 
 .container {
 	min-height: 324px;
