+++ src/2_widgets/Definitions/DefinitionsList/item/DefinitionItem.tsx
