+++ src/2_widgets/Definitions/DefinitionsList/item/index.ts
