+++ src/2_widgets/Definitions/DefinitionsWidget/header/DefinitionHeader.module.scss
