+++ src/2_widgets/Definitions/DefinitionsWidget/header/index.ts
