+++ src/2_widgets/Definitions/DefinitionsWidget/index.ts
@@ -1,0 +1,1 @@
+export { DefinitionsWidget } from './DefinitionsWidget';
