+++ src/2_widgets/Definitions/index.ts
@@ -1,0 +1,3 @@
+export { DefinitionsWidget } from './DefinitionsWidget';
+
+export { DefinitionsList } from './DefinitionsList';
