--- src/2_widgets/Definitions/list/DefinitionsList.module.scss
@@ -1,4 +1,0 @@
-.list {
-	display: flex;
-	gap: calc(max(4%, 20px));
-}
